MAJ -- GESTION GRAPHE -- MODIF LIBELLES PERF

+++ client/client-protocol/Data/cpu.h
@@ -21,7 +21,7 @@ public:
 
 private :
     QString nom ;
-    int tauxUtilisation ;
+    double tauxUtilisation ;
     double vitesse;
     double vitesseMax;
     int nombreCoeurs;
